Added a Shaders folder in Assets and an outline shader
This outline shader can be added to a gameobject's material. The outline width can initially be set to 0 and an initial color set (you can also change the colorat runtime). When an enemy has died, the outline width can be changed to say .5 to show the player the enemy is now "possessable".