Create a weapon controller with a basic weapon
I split the ProjectileLauncher script in such a way to extract the
firing logic out of the launcher script. That way the WeaponController
will be in the player, and the IWeapon will be in a gameobject within
the player, which can be swapped out during possession. This also allows
enemies to use the same firing logic without having to alter it
specially for them.